requirements.txt
torch>=1.12.0
datasets>=2.0.0
tokenizers>=0.12.0
numpy>=1.21.0
PyYAML>=6.0
matplotlib>=3.5.0